smarttask_gptneo/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
transformers==4.33.2
torch==2.0.1
schedule==1.2.0
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0